1)Cash summ 2) application Icon 3) Release application

+++ app/src/main/java/com/example/housemanagment/models/blockData/Block.kt
@@ -7,5 +7,5 @@ data class Block(
     val dom_count: Int,
     val id: Int,
     val name: String,
-    val summa: Int
+    val summa: Double
 ):Serializable
