+++ app/src/main/java/com/example/housemanagment/models/buildingData/Building.kt
@@ -7,5 +7,5 @@ data class Building(
     val blok_count: Int,
     val id: Int,
     val name: String,
-    val summa: Int
+    val summa: Double
 ):Serializable
