+++ app/src/main/java/com/example/housemanagment/models/flat/Flat.kt
@@ -8,5 +8,5 @@ data class Flat(
     val home_count: Int,
     val id: Int,
     val name: String,
-    val summa: Int
+    val summa: Double
 ):Serializable
